refactor and update for custom fim templates
